Alembic procedural: deduplicate cached data accross frames
Currently the procedural will add an entry to the cache for every frame
even if the data only changes seldomly. This means that in some cases we
will have duplicate data accross frames.

Now we compute a hash for the data and compare it to the one for the
previous frame and only if they are different do we create a new entry.
The hash is computed by the Alembic library (using murmur3).

This can drastically reduce memory usage at the cost of more processing
time. In one file, memory usage goes from 9.08 Gb down to 4.28 Gb over
240 frames.