Fix a few issues with prefetching.
We need to make sure that the data for the frame that we are trying to
render is indeed in the prefetched cache if it is not in the main cache.

We also need to avoid race conditions when reading from a partially
loaded cache.

Finally, use a specific cancel query mechanism to early exit when
cancelling prefetching.