Support "Release Confirm" mode for the pose blending operator
This will make it possible to click and drag on the poses in the
asset view template to do the blending. On mouse release, the blended
pose will be applied.
This isn't used yet, follow-up commits will enable it.

This also hides the cursor and enables continuous grabbing when dragging
with the "Release Confirm" enabled. To be evaluated; maybe it should
always do this while blending.

Also fixes mouse hover highlights not being updated when confirming or
cancelling the pose blending operator.