Expose more asset info and utilities in the Python API
Needed so the pose library add-on can keep its operators work outside of
the Asset Browser context, namely with the asset view template.

An issue here is that we can't expose an asset handle from the Asset
Browser, because there is not permanently saved one whose pointer we can
pass around. This is planned to be addressed, but for now Python will
have to use the file wrapped by the asset-handle directly, via
`context.asset_file_handle`.