GPencil: Fix bezier stroke weight painting
The last element of a bezier stroke could not be weight painted.
This was because the checks use line segments to test for a hit and then
pick the first point of that segment to paint the weight. This causes the
problems for the last point.

Instead of testing of the segment is inside the circle, we just test
if the points are inside the circle.

Note that this should also be done for poly strokes at some point,
because it makes the code simpler.