GPencil: Add find_cubic_bezier_stationary_points
This function calculates the stationary points of a bezier curve.

The new fitting implementation will use this function to determin where
to place control points so that attributes, like the thickness, can be
fitted to the curve. Since we do not have handles to control the falloff
of attributes, we have to find points of interest where our
interpolation between two points gives good results.