USD import: root object xform logic fix.
Fixed logic for determining if a USDXformReader object
is the root of a transform hierarchy.  The previous
implementation did not work for USDXformReader instances
created for transform cache contraints, since in that case
parent readers aren't set. Also, now checking if the
prim is a root xform in the USDXformReader constructor,
to avoid repeating this check every time the matrix
is computed.