Apply D9974: Sanitize threaded preview creation with undo
With undo/redo after creating assets, previews of the "Current File"
asset library would often go missing, e.g. see T83884.

The preview generation in a background job will eventually modify ID
data, but the undo push was done prior to that. So obviously, an undo
then would mean the preview is lost.

This patch makes it so undo/redo will regenerate the preview, if the
preview rendering was invoked but not finished in the undone/redone
state.

The preview flag `PRV_UNFINISHED` wasn't entirely what we needed. So I
had to change it to a slightly different flag.