Show asset path in asset browser sidebar + Open Blend File button
This is useful information to have and the Open Blend File makes sense to have
next to it.

A general Asset Browser feature really, but the Open Blend File operator is
only available in this branch, so committing that here.

Also added a TODO note about something that would be especially good to have
for this change.