Resolve function to set runtime type details.
This is optional node function can be used to resolve type details such
as base type, attribute domain, etc. at runtime. It is intended for
nodes whose operation depends on the concrete runtime type of outputs.
It gets called in a back-propagating manner so all outputs are resolved
before the function is called on a node.