Fix: Unavailable sockets are used during compositing execution
When translating node tree into operations, unavailable sockets (with flag SOCK_UNAVAIL) are converted into operations sockets too. This causes an assertion to fail when selecting a multi-layer image with a "Combined" layer in image node because it maps associated "Alpha" socket to output operation socket twice (as there is an "Image" socket too but with SOCK_UNAVAIL flag).

ImageNode::convertToOperations assumes there is a single socket for combined  pass but in this case the unavailable "Image" socket is used too for combined pass. 
Image node sockets are marked with SOCK_UNAVAIL flag instead of being removed to keep connections of unused sockets but during compositing execution they shouldn't be created as they are not executed.
This problem has become visible after commit 93e2491, as now map add_new is used instead of just overwriting last mapped socket.