Compositor: Crop operations canvases to rendered areas
When calculating operations resolutions (canvas size), rendered areas are not known yet so worst case scenario is assumed which is full size. But they can be cropped to rendered areas bounds one they are known. This reduces memory usage as buffers size will be the same than rendered areas bounds.