Compositor: Fix image node alpha socket converted into operations twice
When translating image node sockets into operations, unavailable ones (with flag SOCK_UNAVAIL) are converted too. This causes an assertion to fail when selecting a multi-layer image with a "Combined" layer because it's mapping associated "Alpha" socket to operation output twice (as there is an unavailable "Image" socket used for combined too).

During compositing execution there is not need to create unavailable sockets as they are not linked or executed. Assertion fails since commit 93e2491, as now map add_new is used instead of just overwriting last mapped socket.