WIP: Geometry Nodes: Support nodes modifier on curve objects
With this commit, curve objects support the nodes modifier. The goal is
that nodes modifiers will read curve data as the original curve, unless
they are placed after a modifier that must tessalate the curve (convert
it to a mesh). Always converting to a mesh in that case  is basically
just a formalization of the behavior that was already present after
recent fixes for 2.8 versions last year and two years ago. For more
controlled explicit conversions between data types, using geometry
nodes makes so much more sense anyway.

TODO:
 - Add more code comments
 - Fix memory leak(s?)
 - Fix geometry components besides mesh don't display in viewport.

Differential Revision: https://developer.blender.org/D11597