Raycast geometry node.
The *Raycast* node intersects rays from one geometry onto another.
It computes hit points on the target mesh and returns normals, distances
and any surface attribute specified by the user.

A ray starts on each point of the input *Geometry*. Rays continue
in the *Ray Direction* until they either hit the *Target Geometry*
or reach the *Ray Length* limit. If the target is hit, the value of the
*Is Hit* attribute in the output mesh will be true. *Hit Position*,
*Hit Normal*, *Hit Distance* and *Hit Index* are the properties of the
target mesh at the intersection point. In addition, a *Target Attribute*
can be specified that is interpolated at the hit point and the result
stored in *Hit Attribute*.

Differential Revision: https://developer.blender.org/D11619