VSE strip thumbnail improvements
The left handle now slides over the thumbs when moving and updates the thumb
under the left handle to represent seq->startdisp. Strip slip offset has fixed
thumbs and they don't update images as they move but rather move along with the
strip itself. Helps to make the strip feel like it is a container for the movie
and the handle areas cut the movie to the strip size. Better UX.