USDStageReader refactor.
Now passing the import params and settings as arguments
to the constructor, to reduce the number of arguments
required to invoke USDStageReader::collect_readers().
The constructor now takes a stage pointer as an argument,
to simplify error checking and initalization of the
import settings. Miscellaneous additional cleanup to
simplify the code.