Refactor: Move vertex group names to object data (WIP)
As explained in T88951, the list of vertex group names is currently
stored separately per object, even though vertex group data is stored
on the geometry. This tends to complicate code and cause bugs,
especially as geometry is created procedurally and tied less closely
to objects.

This commit moves the storage of `bDeformGroup` to `Mesh`, `Lattice`,
and `bGPdata` instead of `Object`. In order to avoid dramatically
changing existing code, I introduced a helper function to access the
list of names given just the object.

TODO:
 - Move the active group index to the object as well.
 - Possibly move the Python API for vertex groups to object data.

Differential Revision: https://developer.blender.org/D11689