Knife: Added local constrained angle mode
When using constrained angle mode, pressing C once will bring you to the current screen-space mode and pressing it again will now enter a new "local" mode.

This mode will snap along the plane of the current cut segment's face.
In the case where the cut was started on a edge or vertex it will snap along the plane of the current face the mouse is over, if this face is adjacent to that edge or vertex.

It made sense to implement a mode that works in this way after adding visible distance and angle measurements as now a user can snap the visible angle for the current cut.