Knife: Removed 'E' knife cancel functionality and moved 'M' modal to 'D'
It is dangerous to have 'E' be able to cancel a user's entire cut as it is a key many could accidentally hit and there is no way to retrieve a lost cut. Removing it leaves 'Esc' as the key for cancelling a cut.

On many keyboards 'M' is far away from the other modal keys for the knife tool. 'D' allows for a better workflow by being closer.